fix: force new talosconfig if endpoints or nodes change
This PR fixes a bug where terraform was unable to apply when the number
of control plane nodes was scaled from, for example, 1 to 3. This
ensures that the talosconfig resource gets recreated any time the IPs in
the list of endpoints or nodes is changed and that gets bubbled out to
all of the other dependent objects. Seems to work well, tested with DO
example.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>